Tests/LibWeb/Text/expected/UIEvents/KeyEvent-ctrl-c.txt

key="c" code=KeyC ctrlKey=true
key="C" code=KeyC ctrlKey=true
